+++ app-backend/db/src/main/scala/ProjectDao.scala
@@ -630,15 +630,12 @@ object ProjectDao
         infoList.length match {
           case 0 => None
           case _ =>
-            val (property, taskType): (List[String], String) =
-              infoList.traverse(_.labelGroupName) match {
-                case Some(groupNameList) =>
-                  (groupNameList.distinct, "classification")
-                case _ => (List("label"), "detection")
-              }
             val stacClasses
               : List[StacLabelItemProperties.StacLabelItemClasses] = (infoList
-              .groupBy(_.labelGroupName.getOrElse("label"))
+              .groupBy(_.labelGroupName match {
+                case Some(groupName) if groupName.nonEmpty => groupName
+                case _                                     => "label"
+              })
               .map {
                 case (propName, info) =>
                   StacLabelItemProperties.StacLabelItemClasses(
@@ -646,6 +643,12 @@ object ProjectDao
                     info.traverse(_.labelName).getOrElse(List())
                 )
             } toList)
+            val groupNamesDry = stacClasses.map(_.name).toSet
+            val (property, taskType) =
+              groupNamesDry.size match {
+                case 1 => (List("label"), "detection")
+                case _ => (groupNamesDry.toList, "classification")
+              }
             Some(
               StacLabelItemPropertiesThin(
                 property,